Update code to allow user input for nth term of Fibonacci series
This commit updates the code to prompt the user for input when running the program. The user can now enter a value for n, which represents the nth term of the Fibonacci series. The program will then generate and display the Fibonacci series up to the nth term.

This change improves the user experience by allowing them to customize the output of the program.